refactor: 룸, 트랙, 미디어소스 역정규화 (#309)
