+++ backend/src/main/java/com/digginroom/digginroom/admin/service/UploadService.java
@@ -1,14 +1,14 @@
 package com.digginroom.digginroom.admin.service;
 
 import com.digginroom.digginroom.admin.controller.dto.UploadRequest;
-import com.digginroom.digginroom.domain.mediasource.MediaSource;
 import com.digginroom.digginroom.domain.room.Room;
 import com.digginroom.digginroom.domain.track.Track;
 import com.digginroom.digginroom.repository.RoomRepository;
 import jakarta.transaction.Transactional;
-import java.util.Arrays;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
 
 @Service
 @RequiredArgsConstructor
@@ -18,7 +18,6 @@ public class UploadService {
 
     @Transactional
     public void save(final UploadRequest request) {
-        MediaSource mediaSource = new MediaSource(request.youtubeVideoId());
         Track track = Track.builder()
                 .title(request.title())
                 .artist(request.artist())
@@ -27,7 +26,7 @@ public class UploadService {
                 .description(request.description())
                 .build();
 
-        Room room = new Room(mediaSource, track);
+        Room room = new Room(request.youtubeVideoId(), track);
         roomRepository.save(room);
     }
 }
