+++ backend/src/main/java/com/digginroom/digginroom/domain/room/Room.java
@@ -1,20 +1,19 @@
 package com.digginroom.digginroom.domain.room;
 
-import static com.digginroom.digginroom.exception.RoomException.NoMediaSourceException;
-
 import com.digginroom.digginroom.domain.BaseEntity;
-import com.digginroom.digginroom.domain.mediasource.MediaSource;
 import com.digginroom.digginroom.domain.track.Track;
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
-import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
+
+import java.util.Objects;
+
+import static com.digginroom.digginroom.exception.RoomException.NoMediaSourceException;
 
 @Getter
 @Entity
@@ -22,23 +21,22 @@ import org.hibernate.annotations.ColumnDefault;
 public class Room extends BaseEntity {
 
     @NotNull
-    @ManyToOne(cascade = CascadeType.PERSIST)
-    private MediaSource mediaSource;
-    @ManyToOne(cascade = CascadeType.PERSIST)
+    private String identifier;
+    @Embedded
     private Track track;
     @Column(nullable = false)
     @ColumnDefault(value = "0")
     private Long scrapCount;
 
-    public Room(final MediaSource mediaSource, final Track track) {
-        validateNotNull(mediaSource);
-        this.mediaSource = mediaSource;
+    public Room(final String identifier, final Track track) {
+        validateNotNull(identifier);
+        this.identifier = identifier;
         this.track = track;
         this.scrapCount = 0L;
     }
 
-    private void validateNotNull(final MediaSource mediaSource) {
-        if (Objects.isNull(mediaSource)) {
+    private void validateNotNull(final String identifier) {
+        if (Objects.isNull(identifier)) {
             throw new NoMediaSourceException();
         }
     }
