+++ backend/src/main/java/com/digginroom/digginroom/domain/track/Track.java
@@ -1,29 +1,19 @@
 package com.digginroom.digginroom.domain.track;
 
-import com.digginroom.digginroom.domain.BaseEntity;
 import com.digginroom.digginroom.domain.Genre;
-import com.digginroom.digginroom.domain.room.Room;
-import jakarta.persistence.Column;
-import jakarta.persistence.Convert;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ThreadLocalRandom;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
+import java.util.List;
+
 @Getter
+@Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(uniqueConstraints = @UniqueConstraint(columnNames = {"title", "artist"}))
-public class Track extends BaseEntity {
+public class Track {
 
     private String title;
     private String artist;
@@ -33,8 +23,6 @@ public class Track extends BaseEntity {
     private List<String> subGenres;
     @Column(length = 500)
     private String description;
-    @OneToMany(mappedBy = "track")
-    private final List<Room> rooms = new ArrayList<>();
 
     @Builder
     public Track(
@@ -49,10 +37,5 @@ public class Track extends BaseEntity {
         this.superGenre = superGenre;
         this.subGenres = subGenres;
         this.description = description;
-    }
-
-    public Room recommendRoom() {
-        int pickedIndex = ThreadLocalRandom.current().nextInt(rooms.size());
-        return rooms.get(pickedIndex);
     }
 }
