--- backend/src/main/java/com/digginroom/digginroom/repository/TrackRepository.java
@@ -1,11 +1,0 @@
-package com.digginroom.digginroom.repository;
-
-import com.digginroom.digginroom.domain.Genre;
-import com.digginroom.digginroom.domain.track.Track;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TrackRepository extends JpaRepository<Track, Long> {
-
-    List<Track> findBySuperGenre(final Genre superGenre);
-}
